Fixing some printing issues and adding a method to change term name

+++ include/barry/model-meat.hpp
@@ -1038,11 +1038,11 @@ inline void Model<Array_Type,Data_Counter_Type,Data_Rule_Type, Data_Rule_Dyn_Typ
     max_v /= static_cast<int>(nterms() + 1);
     min_v /= static_cast<int>(nterms() + 1);
 
-    printf_barry("Num. of Arrays       : %li\n", this->size());
-    printf_barry("Support size         : %li\n", this->size_unique());
+    printf_barry("Num. of Arrays       : %i\n", static_cast<int>(this->size()));
+    printf_barry("Support size         : %i\n", static_cast<int>(this->size_unique()));
     printf_barry("Support size range   : [%i, %i]\n", min_v, max_v);
     printf_barry("Transform. Fun.      : %s\n", transform_model_fun ? "yes": "no");
-    printf_barry("Model terms (%li)    :\n", this->nterms());
+    printf_barry("Model terms (%i)    :\n", static_cast<int>(this->nterms()));
     for (auto & cn : this->colnames())
     {
         printf_barry(" - %s\n", cn.c_str());
@@ -1050,7 +1050,10 @@ inline void Model<Array_Type,Data_Counter_Type,Data_Rule_Type, Data_Rule_Dyn_Typ
 
     if (this->nrules() > 0u)
     {
-        printf_barry("Model rules (%li)     :\n", this->nrules());
+        printf_barry(
+            "Model rules (%i)     :\n",
+            static_cast<int>(this->nrules())
+        );
     
         for (auto & rn : rules->get_names())
         {
@@ -1060,7 +1063,10 @@ inline void Model<Array_Type,Data_Counter_Type,Data_Rule_Type, Data_Rule_Dyn_Typ
 
     if (this->nrules_dyn() > 0u)
     {
-        printf_barry("Model rules dyn (%li):\n", this->nrules_dyn());
+        printf_barry(
+            "Model rules dyn (%i):\n",
+            static_cast<int>(this->nrules_dyn())
+        );
     
         for (auto & rn : rules_dyn->get_names())
         {
